Uses a graph to parse data from a movie dataset
Takes the IMBDDataset.tsv file and parses it to a list, then places the actors/actresses from each movie as a node on an unweighted graph. The edges connecting the nodes show that the actors/actresses starred in the same movie. Then parses the graph to find the person with the highest degree and shows the degree of a single actress.